refactor TextureUtils into sub-modules
# Conflicts:
#	scripts/Engineering/texture/TextureUtils/fitting_utils.py
