+++ scripts/Engineering/texture/TextureUtils/__init__.py
@@ -1,0 +1,34 @@
+# Maintain the API for importing through the TextureUtils package
+
+from .io import find_all_files, mk
+from .focus_utils import run_focus_script
+from .correction_utils import run_abs_corr, validate_abs_corr_inputs
+from .fitting_utils import (
+    get_initial_fit_function_and_kwargs_from_specs,
+    fit_initial_summed_spectra,
+    rerun_fit_with_new_ws,
+    crop_and_rebin,
+    crop_wss_and_combine,
+    calc_intens_and_sigma_arrays,
+    fit_all_peaks,
+)
+from .polefigure_utils import create_pf, create_pf_loop, make_iterable, plot_pole_figure
+
+__all__ = [
+    find_all_files,
+    mk,
+    run_focus_script,
+    run_abs_corr,
+    validate_abs_corr_inputs,
+    get_initial_fit_function_and_kwargs_from_specs,
+    fit_initial_summed_spectra,
+    rerun_fit_with_new_ws,
+    crop_and_rebin,
+    crop_wss_and_combine,
+    calc_intens_and_sigma_arrays,
+    fit_all_peaks,
+    create_pf,
+    create_pf_loop,
+    make_iterable,
+    plot_pole_figure,
+]
