+++ scripts/Engineering/texture/TextureUtils/fitting_utils.py
@@ -1,291 +1,25 @@
 # Mantid Repository : https://github.com/mantidproject/mantid
 #
-# Copyright &copy; 2025 ISIS Rutherford Appleton Laboratory UKRI,
+# Copyright &copy; 2026 ISIS Rutherford Appleton Laboratory UKRI,
 #   NScD Oak Ridge National Laboratory, European Spallation Source,
 #   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 # SPDX - License - Identifier: GPL - 3.0 +
+
 import numpy as np
-from os import path, scandir
-from Engineering.texture.correction.correction_model import TextureCorrectionModel
-from Engineering.texture.polefigure.polefigure_model import TextureProjection
+from os import path
 from mantid.simpleapi import SaveNexus, logger, CreateEmptyTableWorkspace, Fit
 from mantid.simpleapi import ConvertUnits, Rebunch, Rebin, SumSpectra, AppendSpectra, CloneWorkspace, CropWorkspaceRagged, Load
 from pathlib import Path
 from Engineering.EnginX import EnginX
 from Engineering.IMAT import IMAT
 from mantid.api import AnalysisDataService as ADS, MultiDomainFunction, FunctionFactory
-from typing import Optional, Sequence, Union, Tuple
+from typing import Optional, Sequence, Tuple
 from mantid.dataobjects import Workspace2D
 from plugins.algorithms.IntegratePeaks1DProfile import get_eval_ws, calc_sigma_from_summation
-from Engineering.texture.xtal_helper import get_xtal_structure
 from Engineering.EnggUtils import convert_TOFerror_to_derror
-from Engineering.common.instrument_config import get_instr_config
-
-# import texture helper functions so they can be accessed by users through the TextureUtils namespace
-from Engineering.texture.texture_helper import plot_pole_figure
 
 from mantid.kernel import DeltaEModeType, UnitConversion
 from plugins.algorithms.peakdata_utils import PeakData
-
-
-# -------- Utility --------------------------------
-
-
-def _make_composite(peak_func, bg_func):
-    """Build a CompositeFunction from existing C++ function objects without
-    serialising them (which would discard workspace references, fixed-parameter
-    state, etc.).  NumDeriv is NOT set so that each member function uses its own
-    derivative method (analytical for LinearBackground, numerical fallback for
-    IkedaCarpenterPV) — this avoids redundant expensive peak evaluations when
-    computing derivatives for background parameters."""
-    comp = FunctionFactory.createFunction("CompositeFunction")
-    comp.add(peak_func)
-    comp.add(bg_func)
-    return comp
-
-
-def find_all_files(directory):
-    """
-    find all the files in a directory
-
-    directory: directory to iterate over
-    """
-    with scandir(directory) as entries:
-        return [entry.path for entry in entries if entry.is_file()]
-
-
-def mk(dir_path: str):
-    """
-    make a directory
-
-    dir_path: path to make a directory at
-    """
-    p = Path(dir_path)
-    if not p.exists():
-        p.mkdir()
-
-
-# -------- Focus Script Logic--------------------------------
-
-
-def run_focus_script(
-    wss: Sequence[str],
-    focus_dir: str,
-    van_run: str,
-    ceria_run: str,
-    full_instr_calib: str,
-    grouping: Optional[str] = None,
-    prm_path: Optional[str] = None,
-    spectrum_num: Optional[str] = None,
-    groupingfile_path: Optional[str] = None,
-) -> None:
-    """
-    Focus data for use in a texture analysis pipeline. Currently only ENGIN-X is supported,
-    but TextureInstrument class should grow to include others.
-
-    wss: Sequence of workspaces to be focused, can be paths to files or ws names
-    focus_dir: directory of where the focused data should be saved
-    van_run: the run number/ file path of the vanadium calibration run
-    ceria_run: the run number/ file path of the latest ceria calibration run at time of experiment
-    full_instr_calib: path to the full instrument calibration file (can be found in settings of Engineering Diffraction Interface)
-    grouping: key for desired detector grouping, if standard, otherwise use the prm path
-    prm_path: optional path to the grouping prm file (produced during calibration), if using a standard detector grouping,
-              just use the grouping argument
-    spectrum_num: optional string of spectra numbers if desired to define custom grouping by specifying the spectra
-    groupingfile_path: optional path to a grouping ".cal" or ".xml" file, alternative to prm_path
-    """
-    instrument = _get_instrument_from_ws_list(wss)
-    config = get_instr_config(instrument)
-    group = config.group(grouping) if grouping else None
-    match instrument:
-        case "IMAT":
-            TextureInstrument = IMAT
-        case "ENGINX" | "ENGIN-X":
-            TextureInstrument = EnginX
-        case _:
-            raise ValueError(f"Unsupported instrument '{instrument}' for texture focusing. Supported instruments are ENGINX and IMAT.")
-    model = TextureInstrument(
-        vanadium_run=van_run,
-        ceria_run=ceria_run,
-        focus_runs=wss,
-        save_dir=focus_dir,
-        prm_path=prm_path,
-        full_inst_calib_path=full_instr_calib,
-        group=group,
-        spectrum_num=spectrum_num,
-        groupingfile_path=groupingfile_path,
-    )
-
-    mk(focus_dir)
-    model.main()
-
-
-def _get_instrument_from_ws_list(wss):
-    instruments = set()
-    for ws_str in wss:
-        if ADS.doesExist(ws_str):
-            ws = ADS.retrieve(ws_str)
-        else:
-            try:
-                ws = Load(Filename=ws_str)
-            except:
-                logger.error(f"Could not find or load '{ws_str}'")
-                return None
-        instruments.add(ws.getInstrument().getName())
-    instruments = list(instruments)
-    if len(instruments) == 1:
-        return instruments[0]
-    else:
-        logger.error("Workspaces provided have multiple different instruments attached: " + ", ".join(instruments))
-        return None
-
-
-# -------- Absorption Script Logic--------------------------------
-
-
-def run_abs_corr(
-    wss: Sequence[str],
-    ref_ws: Optional[str] = None,
-    orientation_file: Optional[str] = None,
-    orient_file_is_euler: Optional[bool] = None,
-    euler_scheme: Optional[str] = None,
-    euler_axes_sense: Optional[str] = None,
-    copy_ref: bool = False,
-    include_abs_corr: bool = False,
-    monte_carlo_args: Optional[str] = None,
-    gauge_vol_preset: Optional[str] = None,
-    gauge_vol_shape_file: Optional[str] = None,
-    include_atten_table: bool = False,
-    eval_point: Optional[Union[str, float]] = None,
-    eval_units: Optional[str] = None,
-    root_dir: str = ".",
-    include_div_corr: bool = False,
-    div_hoz: Optional[float] = None,
-    div_vert: Optional[float] = None,
-    det_hoz: Optional[float] = None,
-    clear_ads_after: bool = True,
-) -> None:
-    """
-    Apply absorption correction to data for use in texture analysis pipeline
-    wss: Sequence of workspace names to have corrections calculated and applied
-    ref_ws: Name of the reference workspace, if one required
-    orientation_file: path to the orientation file (should be .txt with one line per run)
-    orient_file_is_euler: flag for whether the file provides euler goniometer angles or direct rotation matrices
-    euler_scheme: the lab frame directions along which each axis of the goniometer initially lies
-    euler_axes_sense: the sense of the rotation around each of the axes (1 being CCW, -1 being CW)
-    copy_ref: Whether the reference sample should be copied to each ws
-    include_abs_corr: Whether the workspaces should have the absorption correction applied
-    monte_carlo_args: String of arguments to supply to the MonteCarloAbsorption alg e.g. "Arg1: val1, Arg2: val2"
-    gauge_vol_preset: Name of the preset to use for the gauge volume, currently ("4mmCube"), otherwise should be Custom or No Gauge Volume
-    gauge_vol_shape_file: Path to custom gauge volume shape file
-    include_atten_table: flag for whether a table of attenuation values at a specified point should be created
-    eval_point: point to calculate the attenuation coefficient at
-    eval_units: units which the eval_point is given in
-    root_dir: Directory path in which the experiment directory is constructed
-    include_div_corr: Flag for whether to include a beam divergence correction
-    div_hoz: Value of beam divergence in the horizontal plane
-    div_vert: Value of beam divergence in the vertical plane
-    det_hoz: Value of divergence on the detector in the horizontal plane
-    clear_ads_after: Flag for whether the produced files should be removed from the ADS after they have been saved
-    """
-    model = TextureCorrectionModel()
-    model.set_reference_ws(ref_ws)
-
-    valid_inputs, error_msg = validate_abs_corr_inputs(
-        ref_ws,
-        orientation_file,
-        orient_file_is_euler,
-        euler_scheme,
-        euler_axes_sense,
-        copy_ref,
-        include_abs_corr,
-        gauge_vol_preset,
-        gauge_vol_shape_file,
-        include_atten_table,
-        eval_point,
-        eval_units,
-        include_div_corr,
-        div_hoz,
-        div_vert,
-        det_hoz,
-    )
-    if not valid_inputs:
-        logger.error(error_msg)
-        return
-    # otherwise run script
-    if orientation_file:
-        model.load_all_orientations(wss, orientation_file, orient_file_is_euler, euler_scheme, euler_axes_sense)
-
-    out_wss = [f"Corrected_{ws}" for ws in wss]
-
-    if copy_ref:
-        model.copy_sample_info(ref_ws, wss)
-
-    model.set_include_abs(include_abs_corr)
-    model.set_include_atten(include_atten_table)
-    model.set_include_div(include_div_corr)
-    model.set_remove_after_processing(clear_ads_after)
-
-    abs_args = {"gauge_vol_preset": gauge_vol_preset, "gauge_vol_file": gauge_vol_shape_file, "mc_param_str": monte_carlo_args}
-
-    atten_args = {"atten_val": eval_point, "atten_units": eval_units}
-
-    div_args = {"hoz": div_hoz, "vert": div_vert, "det_hoz": det_hoz}
-
-    model.calc_all_corrections(wss, out_wss, root_dir=root_dir, abs_args=abs_args, atten_args=atten_args, div_args=div_args)
-
-
-def validate_abs_corr_inputs(
-    ref_ws: Optional[str] = None,
-    orientation_file: Optional[str] = None,
-    orient_file_is_euler: Optional[bool] = None,
-    euler_scheme: Optional[str] = None,
-    euler_axes_sense: Optional[str] = None,
-    copy_ref: bool = False,
-    include_abs_corr: bool = False,
-    gauge_vol_preset: Optional[str] = None,
-    gauge_vol_shape_file: Optional[str] = None,
-    include_atten_table: bool = False,
-    eval_point: Optional[Union[str, float]] = None,
-    eval_units: Optional[str] = None,
-    include_div_corr: bool = False,
-    div_hoz: Optional[float] = None,
-    div_vert: Optional[float] = None,
-    det_hoz: Optional[float] = None,
-) -> Tuple[bool, str]:
-    error_msg = ""
-    # validate inputs
-    if orientation_file:
-        valid_orientation_inputs = isinstance(orient_file_is_euler, bool)
-        if not valid_orientation_inputs:
-            error_msg += r"If orientation file is specified, must flag orient_file_is_euler.\n"
-        if valid_orientation_inputs and orient_file_is_euler:
-            # if is euler flag, require euler_scheme and euler_axes_sense
-            valid_orientation_inputs = isinstance(euler_scheme, str) and isinstance(euler_axes_sense, str)
-            if not valid_orientation_inputs:
-                error_msg += r"If orientation file is euler, must provide scheme and sense.\n"
-
-    if copy_ref:
-        if not isinstance(ref_ws, str):
-            error_msg += r"If copy_ref is True, must provide ref_ws.\n"
-
-    if include_abs_corr:
-        if gauge_vol_preset == "Custom":
-            if not isinstance(gauge_vol_shape_file, str):
-                error_msg += r"If custom gauge volume required, must provide shape xml as file.\n"
-
-    if include_atten_table:
-        if not (isinstance(eval_point, Union[str, float]) and isinstance(eval_units, str)):
-            error_msg += r"If attenuation table required, must provide valid point and units.\n"
-
-    if include_div_corr:
-        if not (isinstance(div_hoz, float) and isinstance(div_vert, float) and isinstance(det_hoz, float)):
-            error_msg += r"If divergence correction required, must provide valid values.\n"
-    # if error_msg is still empty string, the inputs are assumed to be valid
-    return error_msg == "", error_msg
-
-
-# -------- Fitting Script Logic--------------------------------
 
 
 def crop_and_rebin(ws, out_ws, lower, upper, rebin_params):
@@ -309,6 +43,19 @@ def crop_wss_and_combine(wss, peak, lower, upper, output):
         crop_and_rebin(ws, intermediate_ws, lower, upper, rebin_params)
         AppendSpectra(f"rebin_ws_{peak}", intermediate_ws, OutputWorkspace=f"rebin_ws_{peak}")
     return SumSpectra(f"rebin_ws_{peak}", OutputWorkspace=output), cropped_rebinned_wss
+
+
+def _make_composite(peak_func, bg_func):
+    """Build a CompositeFunction from existing C++ function objects without
+    serialising them (which would discard workspace references, fixed-parameter
+    state, etc.).  NumDeriv is NOT set so that each member function uses its own
+    derivative method (analytical for LinearBackground, numerical fallback for
+    IkedaCarpenterPV) — this avoids redundant expensive peak evaluations when
+    computing derivatives for background parameters."""
+    comp = FunctionFactory.createFunction("CompositeFunction")
+    comp.add(peak_func)
+    comp.add(bg_func)
+    return comp
 
 
 def fit_initial_summed_spectra(wss, peaks, peak_window, fit_kwargs, peak_func_name):
@@ -870,228 +617,3 @@ def replace_nans(vals: np.ndarray, method: Optional[str] = None) -> np.ndarray:
     nan_mask = np.isnan(out)
     out[nan_mask] = col_stat[np.where(nan_mask)[1]]
     return out
-
-
-# -------- Pole Figure Script Logic--------------------------------
-
-
-def create_pf(
-    wss: Sequence[str],
-    root_dir: str,
-    exp_name: str,
-    dir1: Sequence[float] = (1.0, 0.0, 0.0),
-    dir2: Sequence[float] = (0.0, 1.0, 0.0),
-    dir3: Sequence[float] = (0.0, 0.0, 1.0),
-    dir_names: Sequence[str] = ("D1", "D2", "D3"),
-    include_scatt_power: bool = False,
-    scatter: bool = True,
-    scat_vol_pos: Sequence[float] = (0.0, 0.0, 0.0),
-    projection_method: str = "Azimuthal",
-    params: Optional[Sequence[str]] = None,
-    xtal_input: Optional[str] = None,
-    xtal_args: Optional[Sequence[str]] = None,
-    hkl: Optional[Sequence[int]] = None,
-    readout_column: Optional[str] = None,
-    kernel: Optional[float] = None,
-    chi2_thresh: Optional[float] = None,
-    peak_thresh: Optional[float] = None,
-    override_dir: bool = False,
-    create_combined_output: bool = False,
-    debug_info_level: int = 0,
-    save_ascii: bool = True,
-) -> None:
-    """
-    Create a single pole figure, for use in texture analysis workflow
-
-    wss: Workspace names of the ws with the orientation information present as a goniometer matrix
-    root_dir: root of the directory to which the data should be saved
-    exp_name: experiment name, which provides the overarching folder within the root directory
-    dir1: vector of the first principle direction of the sample
-    dir2: vector of the second (projection) principle direction of the sample
-    dir3: vector of the third principle direction of the sample
-    dir_names: Names of the first, second and third principle directions
-    include_scatt_power: flag for whether to adjust the value by a scattering power calculation
-    scatter: flag for whether the plotted pole figure should be a scatter plot of experimental points, a fitted contour plot, or both
-    scat_vol_pos: position of the centre of mass of the scattering gauge volume
-    projection_method: the type of projection to use to create the pole figure ("Azimuthal", "Stereographic")
-    params: Parameter Workspaces if you want to read a column of this table to each point in the pole figure
-    xtal_input: method by which the crystal structure will be input, options are ("cif", "array", "string")
-    xtal_args: list of arguments for the specified crystal input:
-                for input "cif", require the cif filepath, example: ["C:/User/Fe.cif",],
-                for "array" array of lattice parameters, space group, basis, example: [(1.0,1.0,1.0), "P1", "Fe 0 0 0 1 0"]
-                for "string" lattice parameter string, space group and basis, example: ["1.0 1.0 1.0", "P1", "Fe 0 0 0 1 0"]
-    hkl: H,K,L reflection of the peak fit in the param workspaces
-    readout_column: column of the param ws that should be attached to the pole figure table
-    kernel: if scatter == False, the kernel size of the gaussian filter applied to smooth the contour plot
-    chi2_thresh: if chi2 column present in params, the maximum value which will still get added to the pole figure table
-    peak_thresh: if X0 present in params, the maximum allowable difference between a spectra's X0 and the mean X0/
-                 X0 corresponding to the provided HKL
-    override_dir: flag which, if True, will save files directly into save_dir rather than creating a folder structure
-    create_combined_output: flag which controls whether to create a combined workspace which contains every spectra in the pole figure
-    debug_info_level: 0 - No debug info; 1 - will label with alpha, beta and value; 2 - will include spectra information in label
-    save_ascii: whether to save files as txt as well as nxs
-    """
-    model = TextureProjection()
-    has_xtal = False
-    if xtal_input:
-        has_xtal = True
-        for ws in wss:
-            ws = ADS.retrieve(ws)
-            xtal = get_xtal_structure(xtal_input, *xtal_args) if xtal_input else None
-            ws.sample().setCrystalStructure(xtal)
-    # only pass the HKL to CreatePoleFigureTable if crystal structure has been defined
-    # otherwise `hkl` is just used for naming the output workspace
-    pf_hkl = hkl if has_xtal else None
-    out_ws, combined_ws, grouping = model.get_pf_output_names(wss, params, hkl, readout_column)
-    # if the flag
-    combined_ws = combined_ws if create_combined_output else None
-
-    dir1, dir2, dir3 = np.asarray(dir1), np.asarray(dir2), np.asarray(dir3)
-    ax_transform = np.concatenate((dir1[:, None], dir2[:, None], dir3[:, None]), axis=1)
-    ax_labels = dir_names
-
-    save_dirs = (
-        [path.join(root_dir, "PoleFigureTables")] if override_dir else model.get_save_dirs(root_dir, "PoleFigureTables", exp_name, grouping)
-    )
-    chi2_thresh = chi2_thresh or 0.0
-    peak_thresh = peak_thresh or 0.0
-    include_spec_info = debug_info_level == 2
-    include_debug_info = debug_info_level in (1, 2)
-    model.make_pole_figure_tables(
-        wss=wss,
-        peak_wss=params,
-        out_ws_name=out_ws,
-        combined_ws_name=combined_ws,
-        save_dirs=save_dirs,
-        hkl=pf_hkl,
-        inc_scatt_corr=include_scatt_power,
-        scat_vol_pos=scat_vol_pos,
-        chi2_thresh=chi2_thresh,
-        peak_thresh=peak_thresh,
-        ax_transform=ax_transform,
-        readout_col=readout_column,
-        include_spec_info=include_spec_info,
-        save_ascii=save_ascii,
-    )
-
-    fig, ax = plot_pole_figure(
-        out_ws,
-        projection_method,
-        fig=None,
-        readout_col=readout_column,
-        save_dirs=save_dirs,
-        plot_exp=scatter,
-        ax_labels=ax_labels,
-        contour_kernel=kernel,
-        display_debug_info=include_debug_info,
-    )
-    ax.set_title(out_ws)
-    try:
-        fig.show()
-    except IndexError:
-        logger.debug("Ignoring a problem with the plt.get_edgecolor. This is (probably) fine")
-
-
-def make_iterable(param):
-    """
-    take a single parameter and make a single value in a list
-
-    param: parameter value
-    """
-    return param if isinstance(param, tuple) or isinstance(param, list) else [param]
-
-
-def create_pf_loop(
-    wss: Sequence[str],
-    param_wss: Sequence[Sequence[str]],
-    include_scatt_power: bool,
-    dir1: Sequence[float],
-    dir2: Sequence[float],
-    dir3: Sequence[float],
-    dir_names: Sequence[str],
-    scatter: Union[str, bool],
-    scat_vol_pos: Sequence[float],
-    save_root: str,
-    exp_name: str,
-    projection_method: str,
-    xtal_input: Optional[str] = None,
-    xtal_args: Optional[Sequence[str]] = None,
-    hkls: Optional[Union[Sequence[Sequence[int]], Sequence[int]]] = None,
-    readout_columns: Optional[Union[str, Sequence[str]]] = None,
-    kernel: Optional[float] = None,
-    chi2_thresh: Optional[float] = None,
-    peak_thresh: Optional[float] = None,
-    create_combined_output: bool = False,
-    debug_info_level: int = 0,
-    save_ascii: bool = True,
-    override_dir: bool = False,
-) -> None:
-    """
-    Create a series of pole figures, for use in texture analysis workflow
-
-    wss: Workspace names of the ws with the orientation information present as a goniometer matrix
-    param_wss: Sequence of Parameter Workspaces if you want to read a column of each table to each point in the pole figure
-    include_scatt_power: flag for whether to adjust the value by a scattering power calculation
-    dir1: vector of the first principle direction of the sample
-    dir2: vector of the second (projection) principle direction of the sample
-    dir3: vector of the third principle direction of the sample
-    dir_names: Names of the first, second and third principle directions
-    scatter: flag as to whether the plotted pole figure should be a scatter plot of experimental points or a fitted contour plot.
-            the string "both" is also a valid argument and that will create both
-    scat_vol_pos: position of the centre of mass of the scattering gauge volume
-    save_root: root of the directory to which the data should be saved
-    exp_name: experiment name, which provides the overarching folder within the root directory
-    projection_method: the type of projection to use to create the pole figure ("Azimuthal", "Stereographic")
-    xtal_input: method by which the crystal structure will be input, options are ("cif", "array", "string")
-    xtal_args: list of arguments for the specified crystal input:
-                for input "cif", require the cif filepath, example: ["C:/User/Fe.cif",],
-                for "array" array of lattice parameters, space group, basis, example: [(1.0,1.0,1.0), "P1", "Fe 0 0 0 1 0"]
-                for "string" lattice parameter string, space group and basis, example: ["1.0 1.0 1.0", "P1", "Fe 0 0 0 1 0"]
-    hkls: H,K,L reflection of each peak fitted by the param workspaces
-    readout_columns: each column of the param ws that should be attached to its own pole figure table
-    kernel: if scatter == False, the kernel size of the gaussian filter applied to smooth the contour plot
-    chi2_thresh: if chi2 column present in params, the maximum value which will still get added to the pole figure table
-    peak_thresh: if X0 present in params, the maximum allowable difference between a spectra's X0 and the mean X0/
-                 X0 corresponding to the provided HKL
-    create_combined_output: flag which controls whether to create a combined workspace which contains every spectra in the pole figure
-    debug_info_level: 0 - No debug info; 1 - will label with alpha, beta and value; 2 - will include spectra information in label
-    save_ascii: whether to save files as txt as well as nxs
-    override_dir: flag which, if True, will save files directly into save_dir rather than creating a folder structure
-    """
-    # get ws paths
-    for iparam, params in enumerate(param_wss):
-        # if multiple peaks are provided, multiple hkls should also be provided
-        hkl = hkls if len(param_wss) == 1 else hkls[iparam] if hkls else None
-
-        for readout_column in make_iterable(readout_columns):
-            kwargs = {
-                "wss": wss,
-                "params": params,
-                "include_scatt_power": include_scatt_power,
-                "xtal_input": xtal_input,
-                "xtal_args": xtal_args,
-                "hkl": hkl,
-                "readout_column": readout_column,
-                "dir1": dir1,
-                "dir2": dir2,
-                "dir3": dir3,
-                "dir_names": dir_names,
-                "kernel": kernel,
-                "scat_vol_pos": scat_vol_pos,
-                "chi2_thresh": chi2_thresh,
-                "peak_thresh": peak_thresh,
-                "root_dir": save_root,
-                "exp_name": exp_name,
-                "projection_method": projection_method,
-                "create_combined_output": create_combined_output,
-                "debug_info_level": debug_info_level,
-                "save_ascii": save_ascii,
-                "override_dir": override_dir,
-            }
-            if scatter == "both":
-                for scat in (True, False):
-                    kwargs["scatter"] = scat
-                    create_pf(**kwargs)
-            else:
-                kwargs["scatter"] = scatter
-                create_pf(**kwargs)
