+++ scripts/test/Engineering/texture/CMakeLists.txt
@@ -1,7 +1,8 @@
 add_subdirectory(correction)
 add_subdirectory(polefigure)
+add_subdirectory(TextureUtils)
 
-set(TEST_PY_FILES test_TextureUtils.py test_xtal_helper.py test_texture_helper.py)
+set(TEST_PY_FILES test_xtal_helper.py test_texture_helper.py)
 
 check_tests_valid(${CMAKE_CURRENT_SOURCE_DIR} ${TEST_PY_FILES})
