+++ scripts/test/Engineering/texture/TextureUtils/test_fitting_utils.py
@@ -9,19 +9,8 @@ import unittest
 from unittest.mock import patch, MagicMock, call
 from os import path
 import numpy as np
-import tempfile
-from Engineering.texture.TextureUtils import (
-    find_all_files,
-    mk,
-    run_focus_script,
-    run_abs_corr,
-    validate_abs_corr_inputs,
-    fit_all_peaks,
+from Engineering.texture.TextureUtils.fitting_utils import (
     get_initial_fit_function_and_kwargs_from_specs,
-    create_pf_loop,
-    _get_run_and_prefix_from_ws_log,
-    _get_grouping_from_ws_log,
-    _get_instrument_from_ws_list,
     fit_initial_summed_spectra,
     rerun_fit_with_new_ws,
     _tie_bkg,
@@ -30,240 +19,12 @@ from Engineering.texture.TextureUtils import (
     _make_composite,
     _get_default_param_ties,
     calc_intens_and_sigma_arrays,
+    _get_run_and_prefix_from_ws_log,
+    _get_grouping_from_ws_log,
+    fit_all_peaks,
 )
-import os
 
-texture_utils_path = "Engineering.texture.TextureUtils"
-
-
-class TextureUtilsFileHelperTests(unittest.TestCase):
-    def test_find_all_files_returns_only_files(self):
-        with tempfile.TemporaryDirectory() as tmpdir:
-            file_path = os.path.join(tmpdir, "file1.txt")
-            dir_path = os.path.join(tmpdir, "subdir")
-            os.mkdir(dir_path)
-            with open(file_path, "w") as f:
-                f.write("test")
-            result = find_all_files(tmpdir)
-            self.assertIn(file_path, result)
-            self.assertNotIn(dir_path, result)
-
-    def test_mk_creates_missing_directory(self):
-        with tempfile.TemporaryDirectory() as tmpdir:
-            test_path = os.path.join(tmpdir, "newdir")
-            self.assertFalse(os.path.exists(test_path))
-            mk(test_path)
-            self.assertTrue(os.path.exists(test_path))
-
-
-class TestGetInstrumentFromWsList(unittest.TestCase):
-    @patch(f"{texture_utils_path}.ADS")
-    def test_single_ws_in_ads_returns_instrument_name(self, mock_ads):
-        mock_ws = MagicMock()
-        mock_ws.getInstrument().getName.return_value = "ENGINX"
-        mock_ads.doesExist.return_value = True
-        mock_ads.retrieve.return_value = mock_ws
-
-        result = _get_instrument_from_ws_list(["ws1"])
-
-        mock_ads.doesExist.assert_called_once_with("ws1")
-        mock_ads.retrieve.assert_called_once_with("ws1")
-        self.assertEqual(result, "ENGINX")
-
-    @patch(f"{texture_utils_path}.ADS")
-    def test_multiple_ws_same_instrument_in_ads_returns_instrument_name(self, mock_ads):
-        mock_ws = MagicMock()
-        mock_ws.getInstrument().getName.return_value = "ENGINX"
-        mock_ads.doesExist.return_value = True
-        mock_ads.retrieve.return_value = mock_ws
-
-        result = _get_instrument_from_ws_list(["ws1", "ws2", "ws3"])
-
-        self.assertEqual(result, "ENGINX")
-        self.assertEqual(mock_ads.retrieve.call_count, 3)
-
-    @patch(f"{texture_utils_path}.Load")
-    @patch(f"{texture_utils_path}.ADS")
-    def test_ws_not_in_ads_loads_from_file(self, mock_ads, mock_load):
-        mock_ws = MagicMock()
-        mock_ws.getInstrument().getName.return_value = "IMAT"
-        mock_ads.doesExist.return_value = False
-        mock_load.return_value = mock_ws
-
-        result = _get_instrument_from_ws_list(["path/to/file.nxs"])
-
-        mock_load.assert_called_once_with(Filename="path/to/file.nxs")
-        self.assertEqual(result, "IMAT")
-
-    @patch(f"{texture_utils_path}.logger")
-    @patch(f"{texture_utils_path}.Load")
-    @patch(f"{texture_utils_path}.ADS")
-    def test_ws_not_loadable_logs_error_and_returns_none(self, mock_ads, mock_load, mock_logger):
-        mock_ads.doesExist.return_value = False
-        mock_load.side_effect = RuntimeError("Cannot load file")
-
-        result = _get_instrument_from_ws_list(["nonexistent.nxs"])
-
-        mock_logger.error.assert_called_once()
-        self.assertIn("nonexistent.nxs", mock_logger.error.call_args[0][0])
-        self.assertIsNone(result)
-
-    @patch(f"{texture_utils_path}.logger")
-    @patch(f"{texture_utils_path}.ADS")
-    def test_multiple_instruments_logs_error_and_returns_none(self, mock_ads, mock_logger):
-        mock_ws1 = MagicMock()
-        mock_ws1.getInstrument().getName.return_value = "ENGINX"
-        mock_ws2 = MagicMock()
-        mock_ws2.getInstrument().getName.return_value = "IMAT"
-        mock_ads.doesExist.return_value = True
-        mock_ads.retrieve.side_effect = [mock_ws1, mock_ws2]
-
-        result = _get_instrument_from_ws_list(["ws1", "ws2"])
-
-        mock_logger.error.assert_called_once()
-        self.assertIn("multiple different instruments", mock_logger.error.call_args[0][0])
-        self.assertIsNone(result)
-
-    @patch(f"{texture_utils_path}.logger")
-    def test_empty_list_logs_error_and_returns_none(self, mock_logger):
-        result = _get_instrument_from_ws_list([])
-
-        mock_logger.error.assert_called_once()
-        self.assertIsNone(result)
-
-    @patch(f"{texture_utils_path}.logger")
-    @patch(f"{texture_utils_path}.Load")
-    @patch(f"{texture_utils_path}.ADS")
-    def test_load_failure_stops_processing_remaining_ws(self, mock_ads, mock_load, mock_logger):
-        # If the second ws fails to load, processing stops early and returns None.
-        mock_ws1 = MagicMock()
-        mock_ws1.getInstrument().getName.return_value = "ENGINX"
-        mock_ads.doesExist.side_effect = [True, False]
-        mock_ads.retrieve.return_value = mock_ws1
-        mock_load.side_effect = RuntimeError("Cannot load")
-
-        result = _get_instrument_from_ws_list(["ws1", "bad_file.nxs"])
-
-        mock_logger.error.assert_called_once()
-        self.assertIn("bad_file.nxs", mock_logger.error.call_args[0][0])
-        self.assertIsNone(result)
-
-
-class TextureUtilsFocusTests(unittest.TestCase):
-    @patch(f"{texture_utils_path}.mk")
-    @patch(f"{texture_utils_path}.IMAT")
-    @patch(f"{texture_utils_path}.EnginX")
-    @patch(f"{texture_utils_path}._get_instrument_from_ws_list", return_value="ENGINX")
-    def test_run_focus_script_instantiates_ENGINX_model_and_calls_main(self, mock_get_instr, mock_enginx, mock_imat, mock_mk):
-        run_focus_script(wss=["1", "2"], focus_dir="focus", van_run="v", ceria_run="c", full_instr_calib="f", grouping="1")
-        mock_enginx.return_value.main.assert_called_once()
-        mock_imat.return_value.main.assert_not_called()
-        mock_mk.assert_called_once_with("focus")
-
-    @patch(f"{texture_utils_path}.mk")
-    @patch(f"{texture_utils_path}.IMAT")
-    @patch(f"{texture_utils_path}.EnginX")
-    @patch(f"{texture_utils_path}._get_instrument_from_ws_list", return_value="IMAT")
-    def test_run_focus_script_instantiates_IMAT_model_and_calls_main(self, mock_get_instr, mock_enginx, mock_imat, mock_mk):
-        run_focus_script(wss=["1", "2"], focus_dir="focus", van_run="v", ceria_run="c", full_instr_calib="f", grouping="1")
-        mock_enginx.return_value.main.assert_not_called()
-        mock_imat.return_value.main.assert_called_once()
-        mock_mk.assert_called_once_with("focus")
-
-
-class TextureUtilsAbsCorrTests(unittest.TestCase):
-    @patch(f"{texture_utils_path}.logger")
-    @patch(f"{texture_utils_path}.validate_abs_corr_inputs")
-    @patch(f"{texture_utils_path}.TextureCorrectionModel")
-    def test_run_abs_corr_executes_full_pipeline(self, mock_model_class, mock_validate, mock_logger):
-        # Arrange
-        mock_model = MagicMock()
-        mock_model_class.return_value = mock_model
-        mock_validate.return_value = (True, "")
-
-        wss = ["ws1", "ws2"]
-        # Act
-        with tempfile.TemporaryDirectory() as d:
-            run_abs_corr(
-                wss=wss,
-                ref_ws="ref",
-                orientation_file="orient.txt",
-                orient_file_is_euler=True,
-                euler_scheme="xyz",
-                euler_axes_sense="right",
-                copy_ref=True,
-                include_abs_corr=True,
-                monte_carlo_args="Arg: Val",
-                gauge_vol_preset="4mmCube",
-                gauge_vol_shape_file="shape.xml",
-                include_atten_table=True,
-                eval_point="1.54",
-                eval_units="Angstrom",
-                root_dir=d,
-                include_div_corr=True,
-                div_hoz=1.0,
-                div_vert=1.0,
-                det_hoz=0.5,
-                clear_ads_after=True,
-            )
-
-        mock_validate.assert_called_once_with(
-            "ref",
-            "orient.txt",
-            True,
-            "xyz",
-            "right",
-            True,
-            True,
-            "4mmCube",
-            "shape.xml",
-            True,
-            "1.54",
-            "Angstrom",
-            True,
-            1.0,
-            1.0,
-            0.5,
-        )
-        mock_logger.error.assert_not_called()
-
-        # Orientation file triggers a load
-        mock_model.load_all_orientations.assert_called_once_with(wss, "orient.txt", True, "xyz", "right")
-
-        # Copy ref when requested
-        mock_model.copy_sample_info.assert_called_once_with("ref", wss)
-
-        # Set processing flags / metadata
-        mock_model.set_include_abs.assert_called_once_with(True)
-        mock_model.set_include_atten.assert_called_once_with(True)
-        mock_model.set_include_div.assert_called_once_with(True)
-        mock_model.set_remove_after_processing.assert_called_once_with(True)
-
-        # corrections call: check args and kwargs
-        mock_model.calc_all_corrections.assert_called_once()
-        args, kwargs = mock_model.calc_all_corrections.call_args
-
-        # Positional args
-        self.assertEqual(args[0], ["ws1", "ws2"])
-        self.assertEqual(args[1], ["Corrected_ws1", "Corrected_ws2"])
-
-        # Keyword args
-        self.assertEqual(
-            kwargs["abs_args"],
-            {"gauge_vol_preset": "4mmCube", "gauge_vol_file": "shape.xml", "mc_param_str": "Arg: Val"},
-        )
-        self.assertEqual(kwargs["atten_args"], {"atten_val": "1.54", "atten_units": "Angstrom"})
-        self.assertEqual(kwargs["div_args"], {"hoz": 1.0, "vert": 1.0, "det_hoz": 0.5})
-
-    @patch(f"{texture_utils_path}.logger")
-    @patch(f"{texture_utils_path}.TextureCorrectionModel")
-    def test_run_abs_corr_logs_error_on_invalid_input(self, mock_model, mock_logger):
-        run_abs_corr(
-            wss=["ws1"],
-            orientation_file="orient.txt",  # Missing orient_file_is_euler
-        )
-        mock_logger.error.assert_called_once()
-        self.assertIn("must flag orient_file_is_euler", mock_logger.error.call_args[0][0])
+texture_utils_path = "Engineering.texture.TextureUtils.fitting_utils"
 
 
 class TextureUtilsFittingUtilsTests(unittest.TestCase):
@@ -1334,78 +1095,6 @@ class TextureUtilsFitAllPeaksNewParamsTests(unittest.TestCase):
         last_call_args = mock_rerun_with_new.call_args_list[-1][0]
         self.assertTrue(last_call_args[-1])  # last_fit_ic
         self.assertTrue(last_call_args[-2])  # is_final
-
-
-class TextureUtilsPoleFigureTests(unittest.TestCase):
-    def get_default_pf_kwargs(self):
-        return {
-            "wss": ["ws1"],
-            "param_wss": [["p1"]],
-            "include_scatt_power": False,
-            "xtal_input": None,
-            "hkls": [[1, 1, 1]],
-            "readout_columns": ["X0", "I"],
-            "dir1": [1, 0, 0],
-            "dir2": [0, 1, 0],
-            "dir3": [0, 0, 1],
-            "dir_names": ["RD", "TD", "ND"],
-            "scatter": "both",
-            "kernel": 0.2,
-            "scat_vol_pos": [0, 0, 0],
-            "chi2_thresh": None,
-            "peak_thresh": None,
-            "save_root": "/save",
-            "exp_name": "exp",
-            "projection_method": "Azimuthal",
-        }
-
-    @patch(f"{texture_utils_path}.create_pf")
-    def test_create_pf_loop_scatter_both_calls_create_twice_per_column(self, mock_create):
-        create_pf_loop(**self.get_default_pf_kwargs())
-        # 2 columns * 2 scatter variants = 4 calls
-        self.assertEqual(mock_create.call_count, 4)
-
-    @patch(f"{texture_utils_path}.create_pf")
-    def test_create_pf_loop_loads_and_calls_create_pf(self, mock_create):
-        with patch(f"{texture_utils_path}.ADS.doesExist", return_value=False):
-            args = self.get_default_pf_kwargs()
-            args["scatter"] = True
-            args["wss"] = ["ws1", "ws2"]
-            args["param_wss"] = (["p1", "p2"],)
-            args["hkls"] = [[1, 1, 1], [1, 0, 0]]
-            args["readout_columns"] = ["I"]
-            create_pf_loop(**args)
-        mock_create.assert_called_once()
-
-    def test_validate_abs_corr_inputs_when_orientation_file_given_requires_is_euler_flag(self):
-        valid, error_msg = validate_abs_corr_inputs(orientation_file="some_file.txt")
-        self.assertFalse(valid)
-        self.assertEqual(error_msg, r"If orientation file is specified, must flag orient_file_is_euler.\n")
-
-    def test_validate_abs_corr_inputs_when_is_euler_require_scheme_and_sense(self):
-        valid, error_msg = validate_abs_corr_inputs(orientation_file="some_file.txt", orient_file_is_euler=True)
-        self.assertFalse(valid)
-        self.assertEqual(error_msg, r"If orientation file is euler, must provide scheme and sense.\n")
-
-    def test_validate_abs_corr_inputs_when_copy_ref_requires_ref_ws(self):
-        valid, error_msg = validate_abs_corr_inputs(copy_ref=True)
-        self.assertFalse(valid)
-        self.assertEqual(error_msg, r"If copy_ref is True, must provide ref_ws.\n")
-
-    def test_validate_abs_corr_inputs_when_include_custom_gv_req_xml_file(self):
-        valid, error_msg = validate_abs_corr_inputs(include_abs_corr=True, gauge_vol_preset="Custom")
-        self.assertFalse(valid)
-        self.assertEqual(error_msg, r"If custom gauge volume required, must provide shape xml as file.\n")
-
-    def test_validate_abs_corr_inputs_when_attenuation_table_req_point_and_unit(self):
-        valid, error_msg = validate_abs_corr_inputs(include_atten_table=True)
-        self.assertFalse(valid)
-        self.assertEqual(error_msg, r"If attenuation table required, must provide valid point and units.\n")
-
-    def test_validate_abs_corr_inputs_when_include_div_corr_require_divergence_vals(self):
-        valid, error_msg = validate_abs_corr_inputs(include_div_corr=True)
-        self.assertFalse(valid)
-        self.assertEqual(error_msg, r"If divergence correction required, must provide valid values.\n")
 
 
 if __name__ == "__main__":
